requirements_production.txt
# Django Framework
Django>=5.2,<5.3

# Django REST Framework pour l'API
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
django-filter>=24.0

# Gestion des variables d'environnement
python-dotenv>=1.0.0

# Base de données (optionnel - pour PostgreSQL en production)
psycopg2-binary>=2.9.9
# OU pour psycopg3:
# psycopg[binary]>=3.1.0

# Cache Redis (optionnel mais recommandé)
django-redis>=5.4.0
redis>=5.0.0

# Gestion des images et QR codes
Pillow>=10.2.0
qrcode[pil]>=7.4.2

# Génération de PDF et rapports
reportlab>=4.0.0
# OU
weasyprint>=60.0

# Export Excel
openpyxl>=3.1.2
xlsxwriter>=3.1.9

# Validation et sécurité
django-environ>=0.11.2

# Monitoring d'erreurs (optionnel mais recommandé en production)
sentry-sdk>=1.40.0

# Tâches asynchrones (optionnel mais recommandé)
celery>=5.3.4
django-celery-beat>=2.5.0

# Serveur WSGI pour production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Tests
pytest>=8.0.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code quality
flake8>=7.0.0
black>=24.1.0
isort>=5.13.0

# Documentation API
drf-yasg>=1.21.7  # Swagger/OpenAPI
# OU
drf-spectacular>=0.27.0

# Utilitaires
python-dateutil>=2.8.2
pytz>=2024.1

# Backup automatique (optionnel)
django-dbbackup>=4.1.0
